Organize test suite and expand URI coverage
What was scattered is now gathered by purpose. The monolithic
test file yields to organized sections—initialization apart from
routing, security separate from utilities. Each test file holds
its proper concern and nothing more.

URI generation gains 27 test cases examining edge conditions,
encoding behaviors, and boundary cases. Not from restless
improvement, but from methodical examination of what should
be tested.

Shared helpers migrate to spec/support where they belong.
Mock applications stand apart from test logic. The structure
serves the tests; the tests serve correctness.

248 examples pass. Organization clarifies without changing
what works. This suffices.
