+++ spec/fixtures/test_util_routes.txt
@@ -1,10 +1,3 @@
 GET / TestApp.index
 GET /show/:id TestApp.show
 POST /create TestApp.create
-PUT /update/:id TestApp.update
-DELETE /delete/:id TestApp.destroy
-GET /error TestApp.error_test
-GET /custom TestApp.custom_headers
-GET /json TestApp.json_response
-GET /html TestApp.html_response
-GET /instance/:id TestInstanceApp#show
